Safari-611.1.10/LayoutTests/accessibility/win/selection-and-focus-expected.txt


This tests that we can adjust focus and selection via accessibility API.


FAIL accessibilityController.focusedElement.role should be document. Was AXRole: AXWebArea.
FAIL accessibilityController.focusedElement.role should be AXList. Was AXRole: AXList.

Option 0 is selected.
Option 1 is not selected.
Option 2 is not selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is selected.
Option 2 is not selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is selected.
Option 2 is selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is not selected.
Option 2 is selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is not selected.
Option 2 is selected.
Option 3 is not selected.

Option 0 is selected.
Option 1 is not selected.
Option 2 is selected.
Option 3 is not selected.

Option 0 is selected.
Option 1 is not selected.
Option 2 is selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is not selected.
Option 2 is selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is not selected.
Option 2 is not selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is not selected.
Option 2 is not selected.
Option 3 is not selected.

== single select ==
Option 0 is selected.
Option 1 is not selected.
Option 2 is not selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is selected.
Option 2 is not selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is selected.
Option 2 is not selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is selected.
Option 2 is not selected.
Option 3 is not selected.

Option 0 is not selected.
Option 1 is selected.
Option 2 is not selected.
Option 3 is not selected.
